Resolve FIXMEs and simplify typing
